--- new-docs/puppeteer.productlauncher.connect.md
@@ -1,22 +1,0 @@
-<!-- Do not edit this file. It is automatically generated by API Documenter. -->
-
-[Home](./index.md) &gt; [puppeteer](./puppeteer.md) &gt; [ProductLauncher](./puppeteer.productlauncher.md) &gt; [connect](./puppeteer.productlauncher.connect.md)
-
-## ProductLauncher.connect() method
-
-<b>Signature:</b>
-
-```typescript
-connect(object: any): any;
-```
-
-## Parameters
-
-|  Parameter | Type | Description |
-|  --- | --- | --- |
-|  object | any |  |
-
-<b>Returns:</b>
-
-any
-
