+++ src/api-docs-entry.ts
@@ -41,9 +41,11 @@ export * from './common/FileChooser.js';
 export * from './common/FrameManager.js';
 export * from './common/Input.js';
 export * from './common/Page.js';
+export * from './common/Product.js';
 export * from './common/Puppeteer.js';
-export * from './node/LaunchOptions.js';
+export * from './common/BrowserConnector.js';
 export * from './node/Launcher.js';
+export * from './node/LaunchOptions.js';
 export * from './common/HTTPRequest.js';
 export * from './common/HTTPResponse.js';
 export * from './common/SecurityDetails.js';
