+++ src/common/Product.ts
@@ -13,30 +13,9 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-/**
- * Launcher options that only apply to Chrome.
- *
- * @public
- */
-export interface ChromeArgOptions {
-  headless?: boolean;
-  args?: string[];
-  userDataDir?: string;
-  devtools?: boolean;
-}
 
 /**
- * Generic launch options that can be passed when launching any browser.
+ * Supported products.
  * @public
  */
-export interface LaunchOptions {
-  executablePath?: string;
-  ignoreDefaultArgs?: boolean | string[];
-  handleSIGINT?: boolean;
-  handleSIGTERM?: boolean;
-  handleSIGHUP?: boolean;
-  timeout?: number;
-  dumpio?: boolean;
-  env?: Record<string, string | undefined>;
-  pipe?: boolean;
-}
+export type Product = 'chrome' | 'firefox';
