+++ src/common/fetch.ts
@@ -13,30 +13,10 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-/**
- * Launcher options that only apply to Chrome.
- *
- * @public
- */
-export interface ChromeArgOptions {
-  headless?: boolean;
-  args?: string[];
-  userDataDir?: string;
-  devtools?: boolean;
-}
 
-/**
- * Generic launch options that can be passed when launching any browser.
- * @public
- */
-export interface LaunchOptions {
-  executablePath?: string;
-  ignoreDefaultArgs?: boolean | string[];
-  handleSIGINT?: boolean;
-  handleSIGTERM?: boolean;
-  handleSIGHUP?: boolean;
-  timeout?: number;
-  dumpio?: boolean;
-  env?: Record<string, string | undefined>;
-  pipe?: boolean;
-}
+import { isNode } from '../environment.js';
+
+/* Use the global version if we're in the browser, else load the node-fetch module. */
+export const getFetch = async (): Promise<typeof fetch> => {
+  return isNode ? await import('node-fetch') : globalThis.fetch;
+};
