+++ src/node/Launcher.ts
@@ -15,29 +15,19 @@
  */
 import * as os from 'os';
 import * as path from 'path';
-import * as http from 'http';
-import * as https from 'https';
-import * as URL from 'url';
 import * as fs from 'fs';
 
 import { BrowserFetcher } from './BrowserFetcher.js';
-import { Connection } from '../common/Connection.js';
 import { Browser } from '../common/Browser.js';
-import { assert } from '../common/assert.js';
-import { debugError } from '../common/helper.js';
-import { ConnectionTransport } from '../common/ConnectionTransport.js';
-import { WebSocketTransport } from '../common/WebSocketTransport.js';
 import { BrowserRunner } from './BrowserRunner.js';
 import { promisify } from 'util';
 
 const mkdtempAsync = promisify(fs.mkdtemp);
 const writeFileAsync = promisify(fs.writeFile);
 
-import {
-  ChromeArgOptions,
-  LaunchOptions,
-  BrowserOptions,
-} from './LaunchOptions.js';
+import { ChromeArgOptions, LaunchOptions } from './LaunchOptions.js';
+import { BrowserOptions } from '../common/BrowserConnector.js';
+import { Product } from '../common/Product.js';
 
 /**
  * Describes a launcher - a class that is able to create and launch a browser instance.
@@ -45,10 +35,9 @@ import {
  */
 export interface ProductLauncher {
   launch(object);
-  connect(object);
   executablePath: () => string;
   defaultArgs(object);
-  product: string;
+  product: Product;
 }
 
 /**
@@ -215,65 +204,8 @@ class ChromeLauncher implements ProductLauncher {
     return resolveExecutablePath(this).executablePath;
   }
 
-  get product(): string {
+  get product(): Product {
     return 'chrome';
-  }
-
-  async connect(
-    options: BrowserOptions & {
-      browserWSEndpoint?: string;
-      browserURL?: string;
-      transport?: ConnectionTransport;
-    }
-  ): Promise<Browser> {
-    const {
-      browserWSEndpoint,
-      browserURL,
-      ignoreHTTPSErrors = false,
-      defaultViewport = { width: 800, height: 600 },
-      transport,
-      slowMo = 0,
-    } = options;
-
-    assert(
-      Number(!!browserWSEndpoint) +
-        Number(!!browserURL) +
-        Number(!!transport) ===
-        1,
-      'Exactly one of browserWSEndpoint, browserURL or transport must be passed to puppeteer.connect'
-    );
-
-    let connection = null;
-    if (transport) {
-      connection = new Connection('', transport, slowMo);
-    } else if (browserWSEndpoint) {
-      const connectionTransport = await WebSocketTransport.create(
-        browserWSEndpoint
-      );
-      connection = new Connection(
-        browserWSEndpoint,
-        connectionTransport,
-        slowMo
-      );
-    } else if (browserURL) {
-      const connectionURL = await getWSEndpoint(browserURL);
-      const connectionTransport = await WebSocketTransport.create(
-        connectionURL
-      );
-      connection = new Connection(connectionURL, connectionTransport, slowMo);
-    }
-
-    const { browserContextIds } = await connection.send(
-      'Target.getBrowserContexts'
-    );
-    return Browser.create(
-      connection,
-      browserContextIds,
-      ignoreHTTPSErrors,
-      defaultViewport,
-      null,
-      () => connection.send('Browser.close').catch(debugError)
-    );
   }
 }
 
@@ -392,63 +324,6 @@ class FirefoxLauncher implements ProductLauncher {
     }
   }
 
-  async connect(
-    options: BrowserOptions & {
-      browserWSEndpoint?: string;
-      browserURL?: string;
-      transport?: ConnectionTransport;
-    }
-  ): Promise<Browser> {
-    const {
-      browserWSEndpoint,
-      browserURL,
-      ignoreHTTPSErrors = false,
-      defaultViewport = { width: 800, height: 600 },
-      transport,
-      slowMo = 0,
-    } = options;
-
-    assert(
-      Number(!!browserWSEndpoint) +
-        Number(!!browserURL) +
-        Number(!!transport) ===
-        1,
-      'Exactly one of browserWSEndpoint, browserURL or transport must be passed to puppeteer.connect'
-    );
-
-    let connection = null;
-    if (transport) {
-      connection = new Connection('', transport, slowMo);
-    } else if (browserWSEndpoint) {
-      const connectionTransport = await WebSocketTransport.create(
-        browserWSEndpoint
-      );
-      connection = new Connection(
-        browserWSEndpoint,
-        connectionTransport,
-        slowMo
-      );
-    } else if (browserURL) {
-      const connectionURL = await getWSEndpoint(browserURL);
-      const connectionTransport = await WebSocketTransport.create(
-        connectionURL
-      );
-      connection = new Connection(connectionURL, connectionTransport, slowMo);
-    }
-
-    const { browserContextIds } = await connection.send(
-      'Target.getBrowserContexts'
-    );
-    return Browser.create(
-      connection,
-      browserContextIds,
-      ignoreHTTPSErrors,
-      defaultViewport,
-      null,
-      () => connection.send('Browser.close').catch(debugError)
-    );
-  }
-
   executablePath(): string {
     return resolveExecutablePath(this).executablePath;
   }
@@ -464,7 +339,7 @@ class FirefoxLauncher implements ProductLauncher {
     }
   }
 
-  get product(): string {
+  get product(): Product {
     return 'firefox';
   }
 
@@ -703,42 +578,6 @@ class FirefoxLauncher implements ProductLauncher {
     );
     return profilePath;
   }
-}
-
-function getWSEndpoint(browserURL: string): Promise<string> {
-  let resolve, reject;
-  const promise = new Promise<string>((res, rej) => {
-    resolve = res;
-    reject = rej;
-  });
-
-  const endpointURL = URL.resolve(browserURL, '/json/version');
-  const protocol = endpointURL.startsWith('https') ? https : http;
-  const requestOptions = Object.assign(URL.parse(endpointURL), {
-    method: 'GET',
-  });
-  const request = protocol.request(requestOptions, (res) => {
-    let data = '';
-    if (res.statusCode !== 200) {
-      // Consume response data to free up memory.
-      res.resume();
-      reject(new Error('HTTP ' + res.statusCode));
-      return;
-    }
-    res.setEncoding('utf8');
-    res.on('data', (chunk) => (data += chunk));
-    res.on('end', () => resolve(JSON.parse(data).webSocketDebuggerUrl));
-  });
-
-  request.on('error', reject);
-  request.end();
-
-  return promise.catch((error) => {
-    error.message =
-      `Failed to fetch browser webSocket url from ${endpointURL}: ` +
-      error.message;
-    throw error;
-  });
 }
 
 function resolveExecutablePath(
